python/widget-store/pyproject.toml
[project]
name = "widget-store"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Peter Kraft", email = "petereliaskraft@gmail.com"},
]
dependencies = [
    "dbos>=0.4.0a0",
    "fastapi[standard]>=0.112.1",
    "python-liquid>=1.12.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
